+++ src/directives/formBuilderDnd.js
@@ -37,6 +37,11 @@ module.exports = [
     $scope.formComponents = formioComponents.components;
     if (!$scope.component) {
       $scope.component = $scope.form;
+      $scope.$watch('form', function(newForm) {
+        if ($scope.component.type === 'form') {
+          $scope.component = newForm;
+        }
+      });
     }
 
     // Components depend on this existing
@@ -50,19 +55,8 @@ module.exports = [
       $scope.$emit.apply($scope, args);
     };
 
-    $scope.$on('iframe-componentClick', function(event, data) {
-      FormioUtils.eachComponent($scope.component.components, function(component) {
-        if (component.id === data.id) {
-          $scope.editComponent(component);
-        }
-      });
-    });
-    $scope.$on('iframe-componentUpdate', function(event, data) {
-      FormioUtils.eachComponent($scope.component.components, function(component) {
-        if (component.id === data.id) {
-          component.overlay = data.overlay;
-        }
-      });
+    $scope.$on('editComponent', function(event, component) {
+      $scope.editComponent(component);
     });
 
     $scope.$on('fbDragDrop', function(event, component) {
@@ -229,7 +223,8 @@ module.exports = [
       }
 
       // Open the dialog.
-      var originalKey = '';
+      $scope.updateKey(component);
+      var originalKey = component.key;
       ngDialog.open({
         template: 'formio/components/settings.html',
         scope: childScope,
@@ -260,9 +255,6 @@ module.exports = [
           // Watch the settings label and auto set the key from it.
           $scope.$watch('component.label', function() {
             $scope.updateKey($scope.component);
-            if (!originalKey) {
-              originalKey = $scope.component.key;
-            }
           });
         }]
       }).closePromise.then(function(e) {
@@ -278,9 +270,9 @@ module.exports = [
         }
 
         // If there is no component label, then set it to the key and set hide label to ensure reverse compatibility.
-        if (!component.label) {
+        // Don't calculate for components that don't have a label.
+        if (!component.label && ['panel', 'content', 'fieldset', 'table', 'well', 'columns'].indexOf(component.type) === -1) {
           component.key = originalKey;
-          $scope.updateKey(component);
           component.label = component.key || component.type;
           component.hideLabel = true;
         }
@@ -288,7 +280,7 @@ module.exports = [
         FormioUtils.eachComponent([component], function(child) {
           delete child.isNew;
         }, true);
-        $scope.$broadcast('iframeMessage', {name: 'updateElement', data: component});
+
         $scope.emit('edit', component);
       });
     };